refactor(sdio): split module code into submodules
